fraud_tracing_engine/requirements.txt
pandas>=2.0.0
networkx>=3.0
numpy>=1.24.0
xgboost>=2.0.0
scikit-learn>=1.3.0
google-cloud-bigquery>=3.0.0

